+++ crates/common/src/commit/request.rs
@@ -4,7 +4,7 @@ use ssz_derive::{Decode, Encode};
 use tree_hash::TreeHash;
 use tree_hash_derive::TreeHash;
 
-use crate::{signature::verify_signed_builder_message, types::Chain, BlstErrorWrapper};
+use crate::{error::BlstErrorWrapper, signature::verify_signed_builder_message, types::Chain};
 
 // TODO: might need to adapt the SignedProxyDelegation so that it goes through
 // web3 signer
