+++ crates/common/src/config.rs
@@ -96,7 +96,7 @@ impl StartSignerConfig {
         let config = CommitBoostConfig::from_env_path()?;
 
         let jwts = load_jwts()?;
-        let server_port = load_env_var_infallible(SIGNER_SERVER_ENV)?.parse()?;
+        let server_port = load_env_var(SIGNER_SERVER_ENV)?.parse()?;
 
         Ok(StartSignerConfig {
             chain: config.chain,
@@ -122,7 +122,7 @@ pub struct ModuleMetricsConfig {
 
 impl ModuleMetricsConfig {
     pub fn load_from_env() -> Result<Self> {
-        let server_port = load_env_var_infallible(METRICS_SERVER_ENV)?.parse()?;
+        let server_port = load_env_var(METRICS_SERVER_ENV)?.parse()?;
         Ok(ModuleMetricsConfig { server_port })
     }
 }
@@ -201,8 +201,8 @@ pub fn load_pbs_custom_config<T: DeserializeOwned>() -> Result<PbsModuleConfig<T
 
     let signer_client = if cb_config.pbs.static_config.with_signer {
         // if custom pbs requires a signer client, load jwt
-        let module_jwt = load_env_var_infallible(MODULE_JWT_ENV)?;
-        let signer_server_address = load_env_var_infallible(SIGNER_SERVER_ENV)?;
+        let module_jwt = load_env_var(MODULE_JWT_ENV)?;
+        let signer_server_address = load_env_var(SIGNER_SERVER_ENV)?;
         Some(SignerClient::new(signer_server_address, &module_jwt))
     } else {
         None
@@ -246,9 +246,9 @@ pub struct StartModuleConfig<T = ()> {
 /// - [MODULE_JWT_ENV] - the jwt token for the module
 // TODO: add metrics url here
 pub fn load_module_config<T: DeserializeOwned>() -> Result<StartModuleConfig<T>> {
-    let module_id = load_env_var_infallible(MODULE_ID_ENV)?;
-    let module_jwt = load_env_var_infallible(MODULE_JWT_ENV)?;
-    let signer_server_address = load_env_var_infallible(SIGNER_SERVER_ENV)?;
+    let module_id = load_env_var(MODULE_ID_ENV)?;
+    let module_jwt = load_env_var(MODULE_JWT_ENV)?;
+    let signer_server_address = load_env_var(SIGNER_SERVER_ENV)?;
 
     #[derive(Debug, Deserialize)]
     struct ThisModuleConfig<U> {
@@ -300,6 +300,6 @@ pub fn load_module_config<T: DeserializeOwned>() -> Result<StartModuleConfig<T>>
     }
 }
 
-pub fn load_env_var_infallible(env: &str) -> Result<String> {
+pub fn load_env_var(env: &str) -> Result<String> {
     std::env::var(env).wrap_err("{env} is not set")
 }
